Change return type of eizo_get_descriptor to int

+++ src/debug.c
@@ -12,15 +12,15 @@ eizo_dbg_dump_descriptor(struct eizo_handle *handle)
 {
     uint8_t desc[HID_MAX_DESCRIPTOR_SIZE];
 
-    ssize_t n = eizo_get_descriptor(handle, desc);
+    int n = eizo_get_descriptor(handle, desc);
     if (n < 0) {
         fprintf(stderr, "%s: reading the descriptor failed.\n", __func__);
         return;
     }
 
-    printf("descriptor size: %ld", n);
+    printf("descriptor size: %d", n);
 
-    for (ssize_t i = 0; i < n; ++i) {
+    for (int i = 0; i < n; ++i) {
         if (i % 16 == 0) {
             printf("\n");
         } else {
